pyproject.toml
[project]
name = "kg"
version = "0.1.0"
description = "Prepare Knowledge Graph of SystemVerilog Codebases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=1.4.0",
    "google-genai>=1.54.0",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "pyverilog>=1.3.0",
    "rdflib>=7.5.0",
]
